:truck: String의 toRule() 확장함수 domain layer로 이동
