+++ data/src/main/java/com/whyranoid/data/model/UserResponse.kt
@@ -1,7 +1,5 @@
 package com.whyranoid.data.model
 
-import com.whyranoid.domain.model.DayOfWeek
-import com.whyranoid.domain.model.Rule
 import com.whyranoid.domain.model.User
 
 data class UserResponse(
@@ -17,12 +15,3 @@ fun UserResponse.toUser() =
         name = this.name,
         profileUrl = this.profileUrl
     )
-
-fun String.toRule(): Rule {
-    val ruleString = this.split("-")
-    return Rule(
-        dayOfWeek = DayOfWeek.values().find { it.dayResId == ruleString[0] } ?: DayOfWeek.SUN,
-        hour = ruleString[1].toInt(),
-        minute = ruleString[2].toInt()
-    )
-}
